backend/requirements.txt
fastapi==0.109.2
uvicorn==0.22.0
SQLAlchemy==2.0.29
alembic==1.12.0
python-multipart==0.0.6
celery==5.3.1
redis==4.5.5
aiofiles==23.1.0
Pillow==10.0.0
psycopg2-binary>=2.9
pydantic[email]==2.12.4
passlib[bcrypt]==1.7.4
python-jose==3.3.0
numpy<2
python-dotenv
requests>=2.31.0
prometheus-fastapi-instrumentator
google-generativeai==0.3.2